Moved andrew testing files to a reference folder and changed outtake motor to a CRServo
