+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/teleOp/testing/ArmAndServo.java
@@ -2,28 +2,35 @@ package org.firstinspires.ftc.teamcode.teleOp.testing;
 
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
+import com.qualcomm.robotcore.hardware.CRServo;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.Servo;
+import com.qualcomm.robotcore.hardware.TouchSensor;
 
+// TODO Change outtake motor to a continuous servo
+// TODO Prevent overextension and underextension of linearSlides
 @TeleOp(name="ArmAndServo")
 public class ArmAndServo extends OpMode {
 
-    double outtakePower = -1.0;
-    double servoRot = 1.0;
+    double outtakePower = 0.0;
+    double armServoRot = 1.0;
     double basketRot = 0.0;
-    boolean clicked = false;
+    boolean clickedX = false;
     boolean claw = false;
     double clawRot = 0.0;
+
+    TouchSensor slideLimit; //slideLimit.isPressed(), assume encoder position is 8000
 
     //Automated time variables
     double current1 = Double.MAX_VALUE;
     double current2 = Double.MAX_VALUE;
     double current3 = Double.MAX_VALUE;
     double current4 = Double.MAX_VALUE;
-    public DcMotor outtakeMotor;
+//    public DcMotor outtakeMotor; Switched to a servo
     public DcMotor vertLinearMotor;
     public DcMotor horizLinearMotor;
 
+    public CRServo outtakeServo; // Goes -1 to 1
     public Servo armServo;
     public Servo clawServo;
     public Servo basketServo;
@@ -33,13 +40,13 @@ public class ArmAndServo extends OpMode {
         telemetry.addData("Initialization","Starting...");
         telemetry.addData("Initialization","Done!");
 
-        outtakeMotor = hardwareMap.dcMotor.get("outtakeMotor"); // Motor Port 0
-        vertLinearMotor = hardwareMap.dcMotor.get("vertLinearMotor"); // Motor Port 1
-        horizLinearMotor = hardwareMap.dcMotor.get("horizLinearMotor"); // Motor Port 2
+        vertLinearMotor = hardwareMap.dcMotor.get("vertLinearMotor"); // Motor Port 0
+        horizLinearMotor = hardwareMap.dcMotor.get("horizLinearMotor"); // Motor Port 1
 
         armServo = hardwareMap.servo.get("rotateServo"); // Servo Port 0
         clawServo = hardwareMap.servo.get("clawServo"); // Servo Port 1
         basketServo = hardwareMap.servo.get("basketServo"); // Servo Port 2
+        outtakeServo = hardwareMap.crservo.get("outtakeServo"); // Servo Port 3
 
         vertLinearMotor.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
     }
@@ -47,17 +54,17 @@ public class ArmAndServo extends OpMode {
     @Override
     public void loop() {
         if (gamepad2.x){
-            clicked = true;
+            clickedX = true;
         }
 
-        if (clicked) {
-            // outtakeMotor (1 is outtake, -1 is intake)
+        if (clickedX) {
+            // outtakeServo (-1 is intake, 1 is outtake)
             // armServo (0 is up, 1 is down)
             if (gamepad2.x) {
                 // Gamepad 1 x pressed -> Stop outtake motor and pull servo up
-                outtakePower = 0;
+                outtakePower = 0.0;
                 current1 = getRuntime();
-                servoRot = 0.0;
+                armServoRot = 0.0;
             }
 
             if (getRuntime() > current1 + .75){
@@ -70,13 +77,13 @@ public class ArmAndServo extends OpMode {
             if (getRuntime() > current2 + 1.5) {
                 // After 1.5 seconds from press -> Stop outtake motor and pull servo down
                 outtakePower = 0.0;
-                servoRot = 1.0;
+                armServoRot = 1.0;
                 current2 = Double.MAX_VALUE;
-                clicked = false;
+                clickedX = false;
             }
         }
         else {
-            outtakePower = gamepad2.right_trigger - gamepad2.left_trigger;
+            outtakePower = (gamepad2.right_trigger) - (gamepad2.left_trigger);
         }
 
         // Basket Drop Servo System
@@ -102,22 +109,20 @@ public class ArmAndServo extends OpMode {
 //
 //        }
 
-        outtakeMotor.setPower(outtakePower);
+        outtakeServo.setPower(outtakePower);
 
         vertLinearMotor.setPower(gamepad2.right_stick_y);
         horizLinearMotor.setPower(gamepad2.left_stick_y);
 
         clawServo.setPosition(clawRot);
-        armServo.setPosition(servoRot);
+        armServo.setPosition(armServoRot);
         basketServo.setPosition(basketRot);
 
-        telemetry.addData("Clicked", clicked);
+        telemetry.addData("Clicked", clickedX);
         telemetry.addData("Claw (open)", claw);
         telemetry.addData("Runtime", getRuntime());
         telemetry.addData("Current1, Current2, Current3", current1 + ", " + current2 + ", " + current3);
         telemetry.addData("g2RStickY, g2LStickY", gamepad2.right_stick_y + ", " + gamepad2.left_stick_y);
         telemetry.update();
-
-
     }
 }
