+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/teleOp/testing/andrewReference/DistanceSensorTest.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.teleOp.testing;
+package org.firstinspires.ftc.teamcode.teleOp.testing.andrewReference;
 
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
